Adjusted the code to handle float values.

+++ main.py
@@ -10,9 +10,9 @@ while str(inputY) == "Y":
     try:
         print("")
         Begindatestring = str(input("Enter date as YYYY-MM-DD: "))
-        g = int(input("Enter the Goal Amount: "))
-        p = int(input("Enter the Payment Amount: "))
-        h = int(input("Enter the Hours per week: "))
+        g = float(input("Enter the Goal Amount: "))
+        p = float(input("Enter the Payment Amount: "))
+        h = float(input("Enter the Hours per week: "))
     except ValueError:
         print("Use the correct values!")
         sys.exit()
